add simple mypy checks and simplify/test custom errors

+++ osmnx/truncate.py
@@ -89,7 +89,7 @@ def truncate_graph_bbox(
     retain_all : bool
         if True, return the entire graph even if it is not connected.
         otherwise, retain only the largest weakly connected component.
-    quadrat_width : numeric
+    quadrat_width : float
         passed on to intersect_index_quadrats: the linear length (in degrees) of
         the quadrats with which to cut up the geometry (default = 0.05, approx
         4km at NYC's latitude)
@@ -136,7 +136,7 @@ def truncate_graph_polygon(
     truncate_by_edge : bool
         if True, retain nodes outside boundary polygon if at least one of
         node's neighbors is within the polygon
-    quadrat_width : numeric
+    quadrat_width : float
         passed on to intersect_index_quadrats: the linear length (in degrees)
         of the quadrats with which to cut up the geometry (default = 0.05,
         approx 4km at NYC's latitude)
